refactor: migrate hooks (use-bulk-operations) from web/app/ce to web/app/core
